requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
asgiref==3.9.1
asttokens==3.0.0
attrs==26.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.6
decorator==5.2.1
Django==5.2.4
django-debug-toolbar==6.2.0
django-phonenumber-field==8.1.0
django-redis==6.0.0
djangorestframework==3.16.1
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.0
frozenlist==1.8.0
gunicorn==25.3.0
http_ece==1.2.1
idna==3.11
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
multidict==6.7.1
packaging==26.0
parso==0.8.5
pexpect==4.9.0
phonenumbers==9.0.13
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg==3.2.9
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-vapid==1.9.4
pycparser==3.0
Pygments==2.19.2
python-decouple==3.8
python-dotenv==1.2.2
pywebpush==2.3.0
qrcode==8.2
redis==7.4.0
requests==2.32.5
sqlparse==0.5.3
stack-data==0.6.3
stripe==13.1.0
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
whitenoise==6.12.0
yarl==1.23.0
